generalize chart rendering

+++ src/main/kotlin/xyz/rthqks/alog/model/AlogDocument.kt
@@ -2,7 +2,7 @@ package xyz.rthqks.alog.model
 
 data class AlogDocument(
     val title: String,
-    val timeStamps: List<Float>,
-    val temp1: List<Float>,
-    val temp2: List<Float>,
+    val timeStamps: List<Double>,
+    val temp1: List<Double>,
+    val temp2: List<Double>,
 )
